requirements.txt
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.12.1
blinker==1.7.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
chromedriver-autoinstaller==0.6.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.1.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
gologin==2023.9.6134110
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
MouseInfo==0.1.3
mss==9.0.1
mypy-extensions==1.0.0
mysql-connector-python==8.2.0
numpy==1.25.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
pathspec==0.12.1
pefile==2023.2.7
Pillow==9.5.0
platformdirs==4.1.0
playwright==1.38.0
protobuf==4.21.12
psutil==5.9.5
py2exe==0.13.0.0
PyAutoGUI==0.9.53
pycparser==2.21
pyee==8.2.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyperclip==1.8.2
pyppeteer==1.0.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
pytz==2023.3
pywhatkit==5.4
requests==2.31.0
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.25
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.8.0
tzdata==2023.3
undetected-chromedriver==3.5.3
urllib3==1.26.16
v==1
webdriver-manager==4.0.1
websockets==10.4
Werkzeug==3.0.1
wikipedia==1.4.0
wsproto==1.2.0
zipp==3.17.0
